chore(release): bump version to v0.4.0
- Cargo.toml 0.3.1 -> 0.4.0, Cargo.lock 同步
- CHANGELOG: 新增 [0.4.0] - 2026-08-31 条目（trace/自适应/P0 8项/M3 5项 + 升级指南）
- docs/installation.md + src/lib.rs 示例 0.3/0.3.1 -> 0.4/0.4.0，版本表新增 0.4.x
- 夹带 clippy 修复：trace.rs redundant closure/to_owned, monitor.rs unnecessary cast, retry.rs dead_code

+++ src/monitor.rs
@@ -166,7 +166,7 @@ impl DownloadMonitor {
                         }
                     }
                     Err(broadcast::error::RecvError::Lagged(skipped)) => {
-                        self.lagged_count += skipped as u64;
+                        self.lagged_count += skipped;
                         ::tracing::warn!(skipped, total_lagged = self.lagged_count, pending_bisects = self.pending_bisects.len(), active_chunks = self.state.chunks.len(), tasks = tasks.len(), "broadcast lagged, skip events");
                         // P0-C1 轻量对账：若 tasks 已空但 state 仍有残留，说明 Complete/Failed 丢失，可能挂死，下次 tick 将尝试重试
                         if tasks.is_empty() && !self.state.chunks.is_empty() {
